requirements.txt
amqp==5.1.1
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
crispy-bootstrap5==0.6
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.5
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-cors-headers==4.1.0
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
django-redis==5.2.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fpdf==1.7.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
kombu==5.2.4
mypy-extensions==1.0.0
packaging==21.3
pathspec==0.11.0
# pdf-creator==1.0.0
Pillow==9.1.1
platformdirs==3.0.0
prompt-toolkit==3.0.29
# psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.6.0
pyparsing==3.0.9
PyPDF2==1.26.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.0a1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
reportlab==4.0.4
requests==2.28.0
rfc3986==1.5.0
sib-api-v3-sdk==7.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tornado==6.2b2
twilio==7.15.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
